backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for the Meat Billing System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic-settings>=2.6.0",
    "pwdlib[argon2]>=0.2.1",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "uvicorn[standard]>=0.32.0",
]
